+++ lib/src/layout/docking_layout.dart
@@ -1,10 +1,17 @@
 import 'package:docking/src/internal/layout/add_item.dart';
+import 'package:docking/src/internal/layout/layout_factory.dart';
 import 'package:docking/src/internal/layout/layout_modifier.dart';
+import 'package:docking/src/internal/layout/layout_stringify.dart';
 import 'package:docking/src/internal/layout/move_item.dart';
 import 'package:docking/src/internal/layout/remove_item.dart';
 import 'package:docking/src/internal/layout/remove_item_by_id.dart';
+import 'package:docking/src/layout/area_builder.dart';
+import 'package:docking/src/layout/docking_area_type.dart';
+import 'package:docking/src/layout/drop_position.dart';
+import 'package:docking/src/layout/layout_parser.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:meta/meta.dart';
 import 'package:multi_split_view/multi_split_view.dart';
 import 'package:tabbed_view/tabbed_view.dart';
 
@@ -13,7 +20,8 @@ mixin DropArea {}
 /// Represents any area of the layout.
 abstract class DockingArea extends Area {
   DockingArea(
-      {double? size,
+      {this.id,
+      double? size,
       double? weight,
       double? minimalWeight,
       double? minimalSize})
@@ -22,6 +30,8 @@ abstract class DockingArea extends Area {
             weight: weight,
             minimalWeight: minimalWeight,
             minimalSize: minimalSize);
+
+  final dynamic id;
 
   int _layoutId = -1;
 
@@ -44,6 +54,11 @@ abstract class DockingArea extends Area {
   bool _disposed = false;
 
   bool get disposed => _disposed;
+
+  Key _key = UniqueKey();
+
+  @internal
+  Key get key => _key;
 
   /// Disposes.
   void _dispose() {
@@ -101,7 +116,7 @@ abstract class DockingArea extends Area {
       throw StateError('Disposed area');
     }
     _parent = parentArea;
-    if (_layoutId != -1 && _layoutId != layoutId) {
+    if (layoutId != -1 && layoutId != layoutId) {
       throw ArgumentError(
           'DockingParentArea already belongs to another layout');
     }
@@ -110,6 +125,7 @@ abstract class DockingArea extends Area {
     return nextIndex;
   }
 
+  /// Converts layout's hierarchical structure to a debug String.
   String hierarchy(
       {bool indexInfo = false,
       bool levelInfo = false,
@@ -131,17 +147,21 @@ abstract class DockingArea extends Area {
     }
     return str;
   }
+
+  String get areaAcronym;
 }
 
 /// Represents an abstract area for a collection of widgets.
 abstract class DockingParentArea extends DockingArea {
   DockingParentArea(List<DockingArea> children,
-      {double? size,
+      {dynamic id,
+      double? size,
       double? weight,
       double? minimalWeight,
       double? minimalSize})
       : this._children = children,
         super(
+            id: id,
             size: size,
             weight: weight,
             minimalWeight: minimalWeight,
@@ -155,9 +175,6 @@ abstract class DockingParentArea extends DockingArea {
         throw ArgumentError('DockingParentArea cannot have disposed child');
       }
     }
-    if (this._children.length < 2) {
-      throw ArgumentError('Insufficient number of children');
-    }
   }
 
   final List<DockingArea> _children;
@@ -167,6 +184,11 @@ abstract class DockingParentArea extends DockingArea {
 
   /// Gets a child for a given index.
   DockingArea childAt(int index) => _children[index];
+
+  /// The first index of [dockingArea] in this container.
+  ///
+  /// Returns -1 if [dockingArea] is not found.
+  int indexOf(DockingArea dockingArea) => _children.indexOf(dockingArea);
 
   /// Whether the [DockingParentArea] contains a child equal to [area].
   bool contains(DockingArea area) {
@@ -188,16 +210,7 @@ abstract class DockingParentArea extends DockingArea {
   @override
   int _updateHierarchy(
       DockingParentArea? parentArea, int nextIndex, int layoutId) {
-    if (this.disposed) {
-      throw StateError('Disposed area');
-    }
-    _parent = parentArea;
-    if (_layoutId != -1 && _layoutId != layoutId) {
-      throw ArgumentError(
-          'DockingParentArea already belongs to another layout');
-    }
-    _layoutId = layoutId;
-    _index = nextIndex++;
+    nextIndex = super._updateHierarchy(parentArea, nextIndex, layoutId);
     for (DockingArea area in _children) {
       nextIndex = area._updateHierarchy(this, nextIndex, layoutId);
     }
@@ -237,7 +250,7 @@ abstract class DockingParentArea extends DockingArea {
 class DockingItem extends DockingArea with DropArea {
   /// Builds a [DockingItem].
   DockingItem(
-      {this.id,
+      {dynamic id,
       this.name,
       required this.widget,
       this.value,
@@ -255,18 +268,19 @@ class DockingItem extends DockingArea with DropArea {
         this.globalKey = keepAlive ? GlobalKey() : null,
         this._maximized = maximized,
         super(
+            id: id,
             size: size,
             weight: weight,
             minimalWeight: minimalWeight,
             minimalSize: minimalSize);
 
-  final dynamic id;
   String? name;
   Widget widget;
   dynamic value;
   bool closable;
   final bool? maximizable;
   List<TabButton>? buttons;
+  @internal
   final GlobalKey? globalKey;
   TabLeadingBuilder? leading;
   bool _maximized;
@@ -298,28 +312,36 @@ class DockingItem extends DockingArea with DropArea {
     }
     return str;
   }
+
+  @override
+  String get areaAcronym => 'I';
 }
 
 /// Represents an area for a collection of widgets.
 /// Children will be arranged horizontally.
 class DockingRow extends DockingParentArea {
   /// Builds a [DockingRow].
-  DockingRow._(List<DockingArea> children,
+  DockingRow._(List<DockingArea> children, dynamic id,
       {double? size,
       double? weight,
       double? minimalWeight,
       double? minimalSize})
       : super(children,
+            id: id,
             size: size,
             weight: weight,
             minimalWeight: minimalWeight,
             minimalSize: minimalSize) {
     controller = MultiSplitViewController(areas: children);
+    if (this._children.length < 2) {
+      throw ArgumentError('Insufficient number of children');
+    }
   }
 
   /// Builds a [DockingRow].
   factory DockingRow(List<DockingArea> children,
-      {double? size,
+      {dynamic id,
+      double? size,
       double? weight,
       double? minimalWeight,
       double? minimalSize}) {
@@ -331,7 +353,7 @@ class DockingRow extends DockingParentArea {
         newChildren.add(child);
       }
     }
-    return DockingRow._(newChildren,
+    return DockingRow._(newChildren, id,
         size: size,
         weight: weight,
         minimalWeight: minimalWeight,
@@ -342,6 +364,9 @@ class DockingRow extends DockingParentArea {
 
   @override
   DockingAreaType get type => DockingAreaType.row;
+
+  @override
+  String get areaAcronym => 'R';
 }
 
 /// Represents an area for a collection of widgets.
@@ -349,21 +374,27 @@ class DockingRow extends DockingParentArea {
 class DockingColumn extends DockingParentArea {
   /// Builds a [DockingColumn].
   DockingColumn._(List<DockingArea> children,
-      {double? size,
+      {dynamic id,
+      double? size,
       double? weight,
       double? minimalWeight,
       double? minimalSize})
       : super(children,
+            id: id,
             size: size,
             weight: weight,
             minimalWeight: minimalWeight,
             minimalSize: minimalSize) {
     controller = MultiSplitViewController(areas: children);
+    if (this._children.length < 2) {
+      throw ArgumentError('Insufficient number of children');
+    }
   }
 
   /// Builds a [DockingColumn].
   factory DockingColumn(List<DockingArea> children,
-      {double? size,
+      {dynamic id,
+      double? size,
       double? weight,
       double? minimalWeight,
       double? minimalSize}) {
@@ -376,6 +407,7 @@ class DockingColumn extends DockingParentArea {
       }
     }
     return DockingColumn._(newChildren,
+        id: id,
         size: size,
         weight: weight,
         minimalWeight: minimalWeight,
@@ -386,6 +418,9 @@ class DockingColumn extends DockingParentArea {
 
   @override
   DockingAreaType get type => DockingAreaType.column;
+
+  @override
+  String get areaAcronym => 'C';
 }
 
 /// Represents an area for a collection of widgets.
@@ -393,7 +428,8 @@ class DockingColumn extends DockingParentArea {
 class DockingTabs extends DockingParentArea with DropArea {
   /// Builds a [DockingTabs].
   DockingTabs(List<DockingItem> children,
-      {bool maximized = false,
+      {dynamic id,
+      bool maximized = false,
       this.maximizable,
       double? size,
       double? weight,
@@ -401,10 +437,15 @@ class DockingTabs extends DockingParentArea with DropArea {
       double? minimalSize})
       : this._maximized = maximized,
         super(children,
+            id: id,
             size: size,
             weight: weight,
             minimalWeight: minimalWeight,
-            minimalSize: minimalSize);
+            minimalSize: minimalSize) {
+    if (this._children.length == 0) {
+      throw ArgumentError('DockingTabs cannot be empty');
+    }
+  }
 
   final bool? maximizable;
 
@@ -418,21 +459,17 @@ class DockingTabs extends DockingParentArea with DropArea {
 
   @override
   void forEach(void f(DockingItem child)) {
-    _children.forEach((element) {
-      f(element as DockingItem);
+    _children.forEach((child) {
+      f(child as DockingItem);
     });
   }
 
   @override
   DockingAreaType get type => DockingAreaType.tabs;
+
+  @override
+  String get areaAcronym => 'T';
 }
-
-/// Represents the [DockingArea] type.
-enum DockingAreaType { item, tabs, row, column }
-
-/// Represents all positions available for a drop event that will
-/// rearrange the layout.
-enum DropPosition { top, bottom, left, right, center }
 
 /// Represents a layout.
 ///
@@ -454,8 +491,18 @@ class DockingLayout extends ChangeNotifier {
   /// The root of this layout.
   DockingArea? get root => _root;
 
+  /// Loads a layout from String.
+  void load(
+      {required String layout,
+      required LayoutParser parser,
+      required AreaBuilder builder}) {
+    root = LayoutFactory.buildRoot(
+        layout: layout, parser: parser, builder: builder);
+  }
+
   /// Set a new root.
   set root(DockingArea? root) {
+    layoutAreas().forEach((area) => area._dispose());
     _root = root;
     _reset();
     notifyListeners();
@@ -489,7 +536,7 @@ class DockingLayout extends ChangeNotifier {
   /// Gets the maximized area in this layout.
   DockingArea? get maximizedArea => _maximizedArea;
 
-  /// Converts layout's hierarchical structure to String.
+  /// Converts layout's hierarchical structure to a debug String.
   String hierarchy(
       {bool indexInfo = false,
       bool levelInfo = false,
@@ -511,19 +558,35 @@ class DockingLayout extends ChangeNotifier {
     _root?._updateHierarchy(null, 1, id);
   }
 
-  /// Finds a [DockingItem] given an id.
-  DockingItem? findDockingItem(dynamic id) {
-    return _findDockingItem(parent: root, id: id);
+  /// Finds a [DockingTabs] that has a [DockingItem] with the given id.
+  DockingTabs? findDockingTabsWithItem(dynamic itemId) {
+    DockingItem? item = findDockingItem(itemId);
+    if (item != null) {
+      DockingArea? parent = item.parent;
+      return parent is DockingTabs ? parent : null;
+    }
+    return null;
   }
 
-  /// Recursively finds a [DockingItem] given an id.
-  DockingItem? _findDockingItem({DockingArea? parent, dynamic id}) {
-    if (parent != null) {
-      if (parent is DockingItem && parent.id == id) {
-        return parent;
-      } else if (parent is DockingParentArea) {
-        for (DockingArea child in parent._children) {
-          DockingItem? item = _findDockingItem(parent: child, id: id);
+  /// Finds a [DockingItem] given an id.
+  DockingItem? findDockingItem(dynamic id) {
+    DockingArea? area = _findDockingArea(area: _root, id: id);
+    return area is DockingItem ? area : null;
+  }
+
+  /// Finds a [DockingArea] given an id.
+  DockingArea? findDockingArea(dynamic id) {
+    return _findDockingArea(area: _root, id: id);
+  }
+
+  /// Recursively finds a [DockingArea] given an id.
+  DockingArea? _findDockingArea({DockingArea? area, dynamic id}) {
+    if (area != null) {
+      if (area.id == id) {
+        return area;
+      } else if (area is DockingParentArea) {
+        for (DockingArea child in area._children) {
+          DockingArea? item = _findDockingArea(area: child, id: id);
           if (item != null) {
             return item;
           }
@@ -581,13 +644,15 @@ class DockingLayout extends ChangeNotifier {
   void moveItem(
       {required DockingItem draggedItem,
       required DropArea targetArea,
-      required DropPosition dropPosition}) {
+      DropPosition? dropPosition,
+      int? dropIndex}) {
     //TODO maximize test
     _rebuild([
       MoveItem(
           draggedItem: draggedItem,
           targetArea: targetArea,
-          dropPosition: dropPosition)
+          dropPosition: dropPosition,
+          dropIndex: dropIndex)
     ]);
   }
 
@@ -609,17 +674,23 @@ class DockingLayout extends ChangeNotifier {
   void addItemOn(
       {required DockingItem newItem,
       required DropArea targetArea,
-      required DropPosition dropPosition}) {
+      DropPosition? dropPosition,
+      int? dropIndex}) {
     //TODO maximize test
     _rebuild([
       AddItem(
-          newItem: newItem, targetArea: targetArea, dropPosition: dropPosition)
+          newItem: newItem,
+          targetArea: targetArea,
+          dropPosition: dropPosition,
+          dropIndex: dropIndex)
     ]);
   }
 
   /// Adds a DockingItem to the root of this layout.
   void addItemOnRoot(
-      {required DockingItem newItem, required DropPosition dropPosition}) {
+      {required DockingItem newItem,
+      DropPosition? dropPosition,
+      int? dropIndex}) {
     if (root == null) {
       throw StateError('Root is null');
     }
@@ -629,7 +700,8 @@ class DockingLayout extends ChangeNotifier {
         AddItem(
             newItem: newItem,
             targetArea: targetArea,
-            dropPosition: dropPosition)
+            dropPosition: dropPosition,
+            dropIndex: dropIndex)
       ]);
     } else {
       throw StateError('Root is not a DropArea');
@@ -683,16 +755,58 @@ class DockingLayout extends ChangeNotifier {
   }
 
   /// Gets recursively all [DockingArea] of that layout.
-  void _fetchAreas(List<DockingArea> areas, DockingArea root) {
-    areas.add(root);
-    if (root is DockingParentArea) {
-      DockingParentArea parentArea = root;
-      parentArea.forEach((child) => _fetchAreas(areas, child));
+  void _fetchAreas(List<DockingArea> areas, DockingArea area) {
+    areas.add(area);
+    if (area is DockingParentArea) {
+      for (DockingArea child in area._children) {
+        _fetchAreas(areas, child);
+      }
     }
   }
 
-  void deserialize(
-      String layout, WidgetDeserialization widgetDeserialization) {}
+  /// Converts a layout into a String to be stored.
+  ///
+  /// The String will have the following structure:
+  /// VERSION:AREAS_LENGTH:AREAS
+  ///
+  /// The VERSION group has a fixed value: V1
+  ///
+  /// The AREAS group separates each area with a comma and follows the
+  /// pattern: AREA_INDEX_1(AREA_CONFIGURATION),...,AREA_INDEX_N(AREA_CONFIGURATION)
+  /// Example: 1(AREA_CONFIGURATION),2(AREA_CONFIGURATION),3(AREA_CONFIGURATION)
+  ///
+  /// The AREA_CONFIGURATION group represents all area data given the class:
+  ///
+  /// * [DockingItem]
+  ///   * AREA_ACRONYM
+  ///   * ID_LENGTH
+  ///   * ID
+  ///   * WEIGHT
+  ///   * MAXIMIZED
+  /// * [DockingColumn]
+  ///   * AREA_ACRONYM
+  ///   * ID_LENGTH
+  ///   * ID
+  ///   * WEIGHT
+  ///   * CHILDREN_INDEXES
+  /// * [DockingRow]
+  ///   * AREA_ACRONYM
+  ///   * ID_LENGTH
+  ///   * ID
+  ///   * WEIGHT
+  ///   * CHILDREN_INDEXES
+  /// * [DockingTabs]
+  ///   * AREA_ACRONYM
+  ///   * ID_LENGTH
+  ///   * ID
+  ///   * WEIGHT
+  ///   * MAXIMIZED
+  ///   * CHILDREN_INDEXES
+  ///
+  /// Example:
+  /// V1:3:1(R;0;;;2,3),2(I;6;my_id1;0.5;F),3(I;6;my_id2;0.5;F)
+  String stringify({required LayoutParser parser}) {
+    final List<DockingArea> areas = layoutAreas();
+    return LayoutStringify.stringify(parser: parser, areas: areas);
+  }
 }
-
-typedef WidgetDeserialization = Widget Function(dynamic itemId);