+++ CHANGELOG.md
@@ -1,4 +1,39 @@
-## 1.12.2
+## 1.15.0
+
+* `DockingLayout`
+  * Method `findDockingTabsWithItem` to finds a `DockingTabs` that has a `DockingItem` with the given id.
+* Theme
+  * The `DockingTheme` class has been added to provide icons customization.
+
+## 1.14.1
+
+* Bugfix
+  * Fixing the issue that reset the weight after using the `DockingLayout.addItemOn` method.
+
+## 1.14.0
+
+* The `id` attribute has been moved from the `DockingItem` to the `DockingArea`.
+* `DockingLayout`
+  * The `findDockingArea` method has been added to locate any area given an id. 
+  * Save/Load layout feature.
+    * The `stringify` and `load` methods have been added to convert the layout to String and to load a layout given a String.
+
+## 1.13.0
+
+* Allow `DockingTabs` to be initialized with only 1 child.
+
+## 1.12.0
+
+* `tabbed_view` dependency updated to 1.18.0
+  * Tab reordering
+* `DropPosition`
+  * The `center` value has been removed.
+* `DockingLayout`
+  * Methods `addItemOnRoot`, `addItemOn` and `moveItem`
+    * New optional parameter: `dropIndex`
+    * The `dropPostion` parameter has become optional.
+
+## 1.11.2
 
 * Bugfix
   * Error with GlobalKey when expanding a DockingTabs.